11/26/2024
fix xong login, đang làm phần farment

+++ src/main/java/com/example/PhongTroOnline/repository/VideoRoomRepository.java
@@ -1,7 +1,8 @@
 package com.example.PhongTroOnline.repository;
 
+import com.example.PhongTroOnline.entity.News;
 import com.example.PhongTroOnline.entity.VideoRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VideoRoomRepository extends JpaRepository<VideoRoom, Integer> {
+public interface VideoRoomRepository extends JpaRepository<VideoRoom, Long> {
 }
